car_rental/requirements.txt
asgiref==3.5.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
factory-boy==3.2.1
Faker==15.1.1
PyJWT==2.6.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
Django==4.1.3
cryptography==38.0.3
